pyREST/requirements.txt
RPI.GPIO
flask
flask-RESTful
